feat: rename telegraf.go to main.go fix windows service issues rename AgentManager to Telegraf
